Refactoring, doc codigo y cambio implementacion agenda, se usa jbdc

+++ src/main/java/com/chiletel/repository/IClienteRepository.java
@@ -1,6 +1,5 @@
 package com.chiletel.repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,8 +7,19 @@ import org.springframework.stereotype.Repository;
 
 import com.chiletel.entity.Cliente;
 
+/**
+ * <h2>Descripción: </h2>
+ * Repositorio de clientes
+ * @author Brayan Baquero
+ */
 @Repository
 public interface IClienteRepository extends JpaRepository<Cliente, Integer>{
 	
-	Optional<Cliente> findByNumeroIden(BigInteger numeroIden);
+	/**
+	 * <h2>Descripción: </h2>
+	 * Buscar a cliente por identificación.
+	 * @param numeroIden
+	 * @return
+	 */
+	Optional<Cliente> findByNumeroIden(long numeroIden);
 }
